+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Basic vgo module added (#136)
  - module/apmhttprouter: added a wrapper type for httprouter.Router to simplify adding routes (#140)
  - Add Transaction.Context methods for setting user IDs (#144)
+ - module/apmgocql: new instrumentation module, providing an observer for gocql (#148)
 
 ## [v0.4.0](https://github.com/elastic/apm-agent-go/releases/tag/v0.4.0)